feat!: use `vim-test` as generic test runner
it's kinda amazing i've made it this far in my career and have been able
to get away with a test runner ONLY for rspec lol.

BREAKING CHANGE: this changes leader mappings to work on (mostly) any
file. functionally it should still run tests but the commands are
completely different now